Updated JUCE, dRow, the VST SDK and made the sources compile again.

+++ projects/SpectrumAnalyser/JuceLibraryCode/modules/juce_core/native/juce_posix_NamedPipe.cpp
@@ -1,30 +1,27 @@
 /*
   ==============================================================================
 
-   This file is part of the juce_core module of the JUCE library.
-   Copyright (c) 2013 - Raw Material Software Ltd.
+   This file is part of the JUCE library.
+   Copyright (c) 2017 - ROLI Ltd.
 
-   Permission to use, copy, modify, and/or distribute this software for any purpose with
-   or without fee is hereby granted, provided that the above copyright notice and this
-   permission notice appear in all copies.
+   JUCE is an open source library subject to commercial or open-source
+   licensing.
 
-   THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH REGARD
-   TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS. IN
-   NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL
-   DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER
-   IN AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN
-   CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+   The code included in this file is provided under the terms of the ISC license
+   http://www.isc.org/downloads/software-support-policy/isc-license. Permission
+   To use, copy, modify, and/or distribute this software for any purpose with or
+   without fee is hereby granted provided that the above copyright notice and
+   this permission notice appear in all copies.
 
-   ------------------------------------------------------------------------------
-
-   NOTE! This permissive ISC license applies ONLY to files within the juce_core module!
-   All other JUCE modules are covered by a dual GPL/commercial license, so if you are
-   using any other modules, be sure to check that you also comply with their license.
-
-   For more details, visit www.juce.com
+   JUCE IS PROVIDED "AS IS" WITHOUT ANY WARRANTY, AND ALL WARRANTIES, WHETHER
+   EXPRESSED OR IMPLIED, INCLUDING MERCHANTABILITY AND FITNESS FOR PURPOSE, ARE
+   DISCLAIMED.
 
   ==============================================================================
 */
+
+namespace juce
+{
 
 class NamedPipe::Pimpl
 {
@@ -33,6 +30,8 @@ public:
        : pipeInName  (pipePath + "_in"),
          pipeOutName (pipePath + "_out"),
          pipeIn (-1), pipeOut (-1),
+         createdFifoIn (false),
+         createdFifoOut (false),
          createdPipe (createPipe),
          stopReadOperation (false)
     {
@@ -47,8 +46,8 @@ public:
 
         if (createdPipe)
         {
-            unlink (pipeInName.toUTF8());
-            unlink (pipeOutName.toUTF8());
+            if (createdFifoIn)  unlink (pipeInName.toUTF8());
+            if (createdFifoOut) unlink (pipeOutName.toUTF8());
         }
     }
 
@@ -119,14 +118,22 @@ public:
         return bytesWritten;
     }
 
-    bool createFifos() const
+    static bool createFifo (const String& name, bool mustNotExist)
     {
-        return (mkfifo (pipeInName .toUTF8(), 0666) == 0 || errno == EEXIST)
-            && (mkfifo (pipeOutName.toUTF8(), 0666) == 0 || errno == EEXIST);
+        return mkfifo (name.toUTF8(), 0666) == 0 || ((! mustNotExist) && errno == EEXIST);
+    }
+
+    bool createFifos (bool mustNotExist)
+    {
+        createdFifoIn  = createFifo (pipeInName, mustNotExist);
+        createdFifoOut = createFifo (pipeOutName, mustNotExist);
+
+        return createdFifoIn && createdFifoOut;
     }
 
     const String pipeInName, pipeOutName;
     int pipeIn, pipeOut;
+    bool createdFifoIn, createdFifoOut;
 
     const bool createdPipe;
     bool stopReadOperation;
@@ -181,14 +188,14 @@ void NamedPipe::close()
 
         char buffer[1] = { 0 };
         ssize_t done = ::write (pimpl->pipeIn, buffer, 1);
-        (void) done;
+        ignoreUnused (done);
 
         ScopedWriteLock sl (lock);
         pimpl = nullptr;
     }
 }
 
-bool NamedPipe::openInternal (const String& pipeName, const bool createPipe)
+bool NamedPipe::openInternal (const String& pipeName, const bool createPipe, bool mustNotExist)
 {
    #if JUCE_IOS
     pimpl = new Pimpl (File::getSpecialLocation (File::tempDirectory)
@@ -202,7 +209,7 @@ bool NamedPipe::openInternal (const String& pipeName, const bool createPipe)
     pimpl = new Pimpl (file, createPipe);
    #endif
 
-    if (createPipe && ! pimpl->createFifos())
+    if (createPipe && ! pimpl->createFifos (mustNotExist))
     {
         pimpl = nullptr;
         return false;
@@ -214,11 +221,13 @@ bool NamedPipe::openInternal (const String& pipeName, const bool createPipe)
 int NamedPipe::read (void* destBuffer, int maxBytesToRead, int timeOutMilliseconds)
 {
     ScopedReadLock sl (lock);
-    return pimpl != nullptr ? pimpl->read (static_cast <char*> (destBuffer), maxBytesToRead, timeOutMilliseconds) : -1;
+    return pimpl != nullptr ? pimpl->read (static_cast<char*> (destBuffer), maxBytesToRead, timeOutMilliseconds) : -1;
 }
 
 int NamedPipe::write (const void* sourceBuffer, int numBytesToWrite, int timeOutMilliseconds)
 {
     ScopedReadLock sl (lock);
-    return pimpl != nullptr ? pimpl->write (static_cast <const char*> (sourceBuffer), numBytesToWrite, timeOutMilliseconds) : -1;
+    return pimpl != nullptr ? pimpl->write (static_cast<const char*> (sourceBuffer), numBytesToWrite, timeOutMilliseconds) : -1;
 }
+
+} // namespace juce
